Applied merged design to current code
